The entire application has been restructured to be more inline with the ModelViewController pattern.
Added loading of multiline FASTA files. However, please note that ONLY RNA FASTA files which contain the letters 'A, C, G, U' can be read. If files which contain the letter 'T' instead of 'U' are loaded, the application will filter out the 'T's, meaning an incomplete sequence will be processed and then visualized!

Added a result bar to the visualization area, which displays a list of the matches found. The result bar is both closable and movable. In addition, a header is shown when displaying the sequence computed. This header can be turned off.

Readded the matrix representation to the application. The matrix representation is now shown in a movable box, which can also be closed if desired.

Readded the dot-bracket representation to the application. The dot-bracket representation can also be turned off.

Improvements have been made to both sequence visualization and circle visualization. These improvements include:

  Both visualizations are now movable and zoomable.
  Matchlines are now colored, and sized more inline with the type of binding.
  Matchlines now overlap less often (sequence only).